Fix `-fix`: replace internal-exposing `Pass` accessor
This is some very weird internal-external code from the Go analysis
package. Because the `Pass` accessor is set by the initialization of
some internal package there, we would have to duplicate the whole
package (and probably its tree of internal dependencies) to use Pass in
checkrunner.

I played with other solutions: a return to using `multichecker`, a
lightly modified `multichecker`, a fork of the go tools repository with
some minor changes... these all sucked.

Until I see a clear issue with the current approach, I'm going to use it
and move on.